Total number of covid deaths and cases in 2020
death_case_ratio is the ratio of the number of deaths and the number of cases. The lower the ratio means there would be fewer people dying from the covid when they are infected, and intuitively, we would assume a country with a lower ratio would have a higher happiness index.